do a better start and initial setup for testsetup.sh
We generate now the ssh-key within the container as a initialization
script, we can now use the openfrontier/docker-gerrit project.

We tell the outside that the service is running via a temporary zero
size file ".started" generated via init script, so we can cycle check
for this file and start the testsetup integration if the file occurs.

Signed-off-by: gerrit-rust test <test@example.com>