topic reviewer fix a remove/add problem
A delete call returnes a erroro or a 204 for empty content. The
deserialze framework returnes a json error code and breaks the removing
of a reviewer.
We destructure now the returned json object and check if status is >=400
for reviewer delete when it is we print an error otherwise a success.

This is a workaround for #17.

Signed-off-by: Silvio Fricke <silvio.fricke@gmail.com>